Refactor deleting previous extra thumbnails in upload-madinah-hotel-gallery.js

+++ middlewares/umrah/packages/upload-madinah-hotel-gallery.js
@@ -36,9 +36,10 @@ module.exports =
         // check if umrah package madinahh hotel extra thumbnails exists
         if (umrahPackage?.madinahhHotelExtraThumbnails?.length > 0) {
             // delete previous extra thumbnails
-            madinahhHotelExtraThumbnails.forEach((thumbnail) => {
-                fs.unlinkSync(path.join(__dirname, '../../../public/', thumbnail));
-            });
+            madinahhHotelExtraThumbnails.forEach(
+                (thumbnail) =>
+                    thumbnail && fs.unlinkSync(path.join(__dirname, '../../../public/', thumbnail))
+            );
         }
 
         // prepare file path
